consumer/requirements.txt
aio-pika
opentelemetry-sdk>=1.25.0
opentelemetry-api>=1.25.0
opentelemetry-exporter-otlp>=1.25.0
opentelemetry-instrumentation-aiohttp-client>=0.50b0
opentelemetry-instrumentation-asyncio>=0.50b0
opentelemetry-instrumentation-logging>=0.50b0
